aysar

+++ routes/userRoutes.js
@@ -155,11 +155,17 @@ router.route('/changeFlightSeats').post((req, res) => {
 router.route('/cancelReservation').post((req, res) => {
   const reservationNumber = (Number(req.body.reservationNumber));
   console.log(Number(reservationNumber))
-  //search if the current user has this reservation number if yes remove it
+  Booking.findOne({reservationNumber: Number(reservationNumber)}).then(ress =>{ 
+    console.log('dfssdfd', req.body.userToken)
+    console.log(ress)
+   console.log('bbb',ress.userID)
+  if(req.body.userToken==ress.userID){
   Booking.findOneAndDelete({reservationNumber:reservationNumber}).then(ress => {
           console.log("cancelled flight", ress)})
-      
-  
+  }else{
+    console.log("cannot cancel this reservation because it belongs to another user")
+  }
+})
 })
 
 router.route('/changePassword').post((req,res)=>{
@@ -179,5 +185,6 @@ router.route('/changePassword').post((req,res)=>{
   
 });
  })
+
 
 export default router;
